Fix Navigation positioning
