+++ package.json
@@ -6,11 +6,11 @@
     "@aspiesoft/random-number-js": "^1.3.2",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.6.3",
+    "@testing-library/user-event": "^12.7.1",
     "gh-pages": "^3.1.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-scripts": "^2.0.5",
+    "react-scripts": "^4.0.2",
     "web-vitals": "^1.1.0"
   },
   "scripts": {
