--- src/App-testing.css
@@ -1,4 +1,0 @@
-.float-top {
-  position: absolute;
-  left: 50%;
-}
