--- src/App.js
@@ -1,275 +1,0 @@
-import React from 'react';
-import conLog from './utils/conLog.js';
-import { Header, Footer } from './HeaderFooter.js';
-
-import { QuizLeftRight as Quiz1 } from './quizzes/QuizLeftRight.js';
-import { QuizLeftMidRight as Quiz2 } from './quizzes/QuizLeftMidRight.js';
-
-import './App.css';
-import './App-testing.css';
-import './flexbox.css';
-
-// Constants
-
-const DEFAULT_QUIZ_ID = 1;
-
-function Main({ currentQuiz }) {
-  conLog(currentQuiz);
-
-  let QuizComponent = null;
-
-  switch (currentQuiz) {
-    case 0:
-    case 1:
-      QuizComponent = Quiz1;
-      break;
-    case 2:
-      QuizComponent = Quiz2;
-      break;
-    default:
-      console.log('Err: no quiz selected.');
-  }
-
-  return (
-    <main className="App-main">
-      <QuizComponent />
-    </main>
-  );
-}
-
-// const DEFAULT_KEY = 0;
-
-function App() {
-  const [currentQuiz, setCurrentQuiz] = React.useState(DEFAULT_QUIZ_ID);
-
-  return (
-    <div className="App">
-      <Header setCurrentQuiz={setCurrentQuiz} />
-      <Main currentQuiz={currentQuiz} />
-      <Footer />
-    </div>
-  );
-}
-
-export default App;
-
-// const DEFAULT_MODE = 'major';
-// const DEFAULT_MODE_ARRAY = 'majorKeys';
-
-// Handy functions
-
-// React Components
-
-// App
-
-// function App() {
-//   const [keyId, updateKeyId] = useState(DEFAULT_KEY);
-//   const [score, updateScore] = useState(0);
-//   const [mode, updateMode] = useState(DEFAULT_MODE);
-//   const [modeArray, updateModeArray] = useState(majorKeys);
-//   const [imgSrc, updateImgSrc] = useState(keyofc);
-//   const [sharpOrFlat, updateSharpOrFlat] = useState('sharp');
-
-//   console.log(`\nApp rerendered:`);
-//   console.log(`\tkeyId: ${keyId}`);
-//   console.log(`\tscore: ${score}`);
-//   console.log(`\tmode: ${mode}`);
-//   // console.log(`\timgSrc: ${imgSrc}`);
-//   // console.log(`\tmodeArray: ${modeArray}`);
-
-//   function formatNoteName(index) {
-//     const noteToFormat = modeArray[index];
-//     // console.log(noteToFormat);
-//     if (noteToFormat) {
-//       let accidental = noteToFormat[2];
-//       if (accidental == 's') {
-//         return `${noteToFormat[0]}♯`;
-//       } else if (accidental == 'f') {
-//         return `${noteToFormat[0]}♭`;
-//       } else return noteToFormat;
-//     }
-//   }
-
-//   function NoteButton({ note, handleClick }) {
-//     const [isWrong, setWrong] = useState(false);
-
-//     const toggleClass = () => {
-//       setWrong(!isWrong);
-//     };
-
-//     let noteLabel = '';
-
-//     const i = parseInt(note, 10);
-
-//     console.log(i);
-
-//     if (modeArray[i].uri == 'choose') {
-//       noteLabel = modeArray[i][sharpOrFlat].label;
-//     } else {
-//       noteLabel = modeArray[i].label;
-//     }
-//     console.log(noteLabel);
-
-//     return (
-//       <div className="tick">
-//         <div className="label">
-//           <button
-//             className={isWrong ? "loser" : null}
-//             onClick={function (e) {
-//               handleClick(note, e);
-//             }}
-//           >
-//             {noteLabel}
-//           </button>
-//         </div>
-//       </div>
-//     );
-//   }
-
-//   function QuestionQuality() {
-//     if (mode == "major") {
-//       return (
-//         <b className="capitalize">{mode}</b>
-//       )
-//       } else {
-//         return <b>{mode}</b>
-//       }
-//   }
-
-//   function generateQuality(accidental) {
-//     let first = 'major';
-//     let second = 'minor';
-//     if (accidental) {
-//       first = 'sharp';
-//       second = 'flat';
-//     } else {
-//       first = 'major';
-//       second = 'minor';
-//     }
-//     return Math.random() >= 0.5 ? first : second;
-//   }
-
-//   function generateNewKey() {
-//     const modeArrayLength = Object.keys(modeArray).length;
-//     let newKeyId = keyId;
-//     let newImgSrc = '';
-
-//     while (newKeyId === keyId) {
-//       const rand = Math.floor(Math.random() * Math.floor(modeArrayLength));
-//       newKeyId = rand;
-//     }
-
-//     const majOrMin = generateQuality();
-//     // console.log(`Your random number is: ${newKeyId} ${majOrMin}`);
-//     updateKeyId(newKeyId);
-
-//     if (majOrMin == 'major') {
-//       updateMode('major');
-//       updateModeArray(majorKeys);
-//       console.log(`\tmajor key:`);
-//     } else {
-//       updateMode('minor');
-//       updateModeArray(minorKeys);
-//       console.log(`\tminor key:}`);
-//     }
-
-//     let newNoteFromArray = {};
-
-//     if (modeArray[newKeyId].uri == 'choose') {
-//       const newSharpOrFlat = generateQuality(true);
-//       if (newSharpOrFlat == 'sharp') {
-//         console.log('generated a sharp key');
-//       } else if (newSharpOrFlat == 'flat') {
-//         console.log('generated a flat key');
-//       } else {
-//         console.log('Error: no sharp or flat given');
-//       }
-//       updateSharpOrFlat(newSharpOrFlat);
-//       newNoteFromArray = modeArray[newKeyId][newSharpOrFlat];
-//     } else {
-//       newNoteFromArray = modeArray[newKeyId];
-//     }
-
-//     // console.log(Object.values(newNoteFromArray));
-
-//     updateImgSrc(newNoteFromArray.uri);
-//   }
-
-//   function handleSkip() {
-//     updateScore(score - 5);
-//     generateNewKey();
-//   }
-
-//   function handleClick(note, e) {
-//     const typeOfNote = typeof note;
-//     const typeOfKeyId = typeof keyId;
-
-//     // console.log(
-//     //   `clicked:\n\t${note}, ${typeOfNote}\nexpected:\n\t${keyId}, ${typeOfKeyId}`
-//     // );
-//     if (parseInt(note, 10) === keyId) {
-//       updateScore(score + 1);
-//       console.log('Winner! Generating a new key.');
-//       generateNewKey();
-//     } else {
-//       updateScore(score - 1);
-//       console.log(e.target);
-//       console.log("YOU'RE A FAILURE, HARRY");
-//     }
-//   }
-
-//   return (
-//     <div className="App">
-//       <header className="App-header">
-//         <div className="brand">
-//           <img src={logo} className="inline App-logo"></img>
-//           <h1 className="inline Logo-title">Flarp!</h1>
-//         </div>
-//       </header>
-//       <main className="App-main">
-//         <div className="child Question-bar">
-//           <h3 className="question">What's the <QuestionQuality /> Key?</h3>
-//         </div>
-//         <div className="child skip-container">
-//           <button className="Skip-button" onClick={handleSkip}>
-//             Skip!
-//           </button>
-//         </div>
-//         <div className="child image-container">
-//           <img src={imgSrc} className="key-img" alt="logo" />
-//         </div>
-//         <div className="child circleOfFifths">
-//           <div className="Score">
-//             <h5>{score}</h5>
-//           </div>
-//           <div className="circle">
-//             <NoteButton note="0" handleClick={handleClick} />
-//             <NoteButton note="1" handleClick={handleClick} />
-//             <NoteButton note="2" handleClick={handleClick} />
-//             <NoteButton note="3" handleClick={handleClick} />
-//             <NoteButton note="4" handleClick={handleClick} />
-//             <NoteButton note="5" handleClick={handleClick} />
-//             <NoteButton note="6" handleClick={handleClick} />
-//             <NoteButton note="7" handleClick={handleClick} />
-//             <NoteButton note="8" handleClick={handleClick} />
-//             <NoteButton note="9" handleClick={handleClick} />
-//             <NoteButton note="10" handleClick={handleClick} />
-//             <NoteButton note="11" handleClick={handleClick} />
-//           </div>
-//         </div>
-//       </main>
-//       <footer className="child App-footer">
-//         <p className="credit">
-//           by <a href="https://github.com/bsantero">BSvdE</a> | <a href="https://github.com/bsantero/flarp-music-quiz">Comments, Suggestions?</a>
-//         </p>
-//         <div className="brand reflection">
-//           <img src={logo} className="inline App-logo"></img>
-//           <h1 className="inline Logo-title">Flarp!</h1>
-//         </div>
-//         {/* <div className="credit-container">
-
-//         </div> */}
-//       </footer>
-//     </div>
-//   );
-// }
