+++ src/components/Footer/Footer.js
@@ -1,23 +1,5 @@
 import React from 'react';
-import Navigation from './Navigation';
-
-import logo from './img/logo.png';
-
-export function Header({ setCurrentQuiz }) {
-  return (
-    <header className="App-header">
-      <div className="brand">
-        <img src={logo} className="inline App-logo" alt="Flarp logo" />
-        <h1 className="inline Logo-title">Flarp!</h1>
-      </div>
-      <Navigation
-        setCurrentQuiz={setCurrentQuiz}
-        className="float-top"
-        label="Switch"
-      />
-    </header>
-  );
-}
+import logo from '../../img/logo.png';
 
 export function Footer() {
   return (
@@ -36,4 +18,4 @@ export function Footer() {
   );
 }
 
-export default { Header, Footer };
+export default { Footer };
