+++ src/components/Navigation/nav-test.css
@@ -1,0 +1,7 @@
+.float-top {
+  width: 100vw;
+  position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
