+++ src/components/Quizzes/QuizLeftRight.js
